style: clippy fixes and reformat

+++ fake/src/faker/impls/barcode.rs
@@ -79,7 +79,7 @@ fn get_properties<L: Data, R: Rng + ?Sized>(_c: L, rng: &mut R) -> IsbnPropertie
     let reg_pub = numerify_sym(&"#".repeat(reg_pub_len), rng);
 
     let mut reg_len = 0;
-    let sufix_reg_pub = &reg_pub[..reg_pub_len as usize - 1];
+    let sufix_reg_pub = &reg_pub[..reg_pub_len - 1];
 
     for r in &rules[group] {
         if r.min <= sufix_reg_pub && sufix_reg_pub <= r.max {
@@ -91,7 +91,7 @@ fn get_properties<L: Data, R: Rng + ?Sized>(_c: L, rng: &mut R) -> IsbnPropertie
         ean,
         group,
         registrant: reg_pub[..reg_len as usize].to_string(),
-        publication: reg_pub[reg_len as usize..reg_pub_len as usize].to_string(),
+        publication: reg_pub[reg_len as usize..reg_pub_len].to_string(),
     }
 }
